Move healthCheck into new package
