+++ src/main/java/requestManagement/fleetManager/healthCheck/HealthCheckImplementation.java
@@ -1,5 +1,6 @@
-package requestManagement.fleetManager;
+package requestManagement.fleetManager.healthCheck;
 
+import requestManagement.fleetManager.FleetManager;
 import requestManagement.fleetManager.hosts.Host;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
